+++ client/src/components/Booking/BookingModal.js
@@ -1,22 +1,14 @@
 import React from 'react';
-import { Form, Field } from 'react-final-form';
 
 import Loader from '../Loader';
-import FormOptions from '../Form/FormOptions';
-import FormErrors from '../Form/FormErrors';
+import BookingModalContent from './BookingModalContent';
 import Modal from '../Modal/Modal';
-import ButtonTwoGroup from '../Button/ButtonTwoGroup';
 
-import { validateForm } from '../../utils/auth';
 import { BOOKINGS_NEED_EVENTS } from '../../const';
 
 import '../Form/Form.css';
 
 const BookingModal = ({ eventOptions, formType, toggleModal, errors, handleOnSubmit }) => {
-    const renderErrors = () => errors.length > 0 && (
-        <FormErrors errors={errors} />
-    );
-
     const renderModalContent = () => {
         if (!eventOptions) {
             return <Loader height={0} />;
@@ -31,37 +23,13 @@ const BookingModal = ({ eventOptions, formType, toggleModal, errors, handleOnSub
         }
 
         return (
-            <Form 
-                validate={(fields) => validateForm(fields, formType)}
-                onSubmit={handleOnSubmit}>
-                {({ handleSubmit }) => (
-                    <div className="form-wrapper">
-                        <form
-                            onSubmit={handleSubmit} 
-                            className="form-container"
-                        >
-                            {renderErrors()}
-                            <Field 
-                                name="event" 
-                                options={eventOptions} 
-                            >
-                                { ({ input, options }) => (
-                                    <FormOptions
-                                        name={input.name}
-                                        options={options}
-                                        onChange={(value) => input.onChange(value)}
-                                        isBooking
-                                    />
-                                )}
-                            </Field>
-                            <ButtonTwoGroup
-                                defaultClass='text-white font-semibold bg-green-400 hover:bg-green-500 transition duration-300'
-                                handleCancel={toggleModal}
-                            />
-                        </form>
-                    </div>
-                )}
-            </Form>
+            <BookingModalContent
+                formType={formType}
+                errors={errors}
+                eventOptions={eventOptions}
+                handleOnSubmit={handleOnSubmit}
+                handleCancelButton={toggleModal}
+            />
         );
     }
